+++ lib/calib/calib_kdf.c
@@ -80,10 +80,10 @@ ATCA_STATUS calib_kdf(ATCADevice device, uint8_t mode, uint16_t key_id, const ui
         packet.param2 = key_id;
 
         // Add details parameter
-        packet.data[0] = details;
-        packet.data[1] = details >> 8;
-        packet.data[2] = details >> 16;
-        packet.data[3] = details >> 24;
+        packet.data[0] = (uint8_t)details;
+        packet.data[1] = (uint8_t)(details >> 8);
+        packet.data[2] = (uint8_t)(details >> 16);
+        packet.data[3] = (uint8_t)(details >> 24);
 
         // Add input message
         if ((mode & KDF_MODE_ALG_MASK) == KDF_MODE_ALG_AES)
