+++ lib/calib/calib_sha.c
@@ -95,7 +95,7 @@ ATCA_STATUS calib_sha_base(ATCADevice device, uint8_t mode, uint16_t length, con
         if (cmd_mode != SHA_MODE_SHA256_PUBLIC && cmd_mode != SHA_MODE_HMAC_START &&
             cmd_mode != SHA_MODE_ECC204_HMAC_START)
         {
-            memcpy(packet.data, message, length);
+            memcpy(packet.data, message, (size_t)length);
         }
 
         if ((status = atSHA(atcab_get_device_type_ext(device), &packet, length)) != ATCA_SUCCESS)
@@ -239,7 +239,7 @@ ATCA_STATUS calib_hw_sha2_256_update(ATCADevice device, atca_sha256_ctx_t* ctx, 
     ATCA_STATUS status = ATCA_SUCCESS;
     uint32_t block_count;
     uint32_t rem_size = ATCA_SHA256_BLOCK_SIZE - ctx->block_size;
-    uint32_t copy_size = data_size > rem_size ? rem_size : (uint32_t)data_size;
+    size_t copy_size = data_size > rem_size ? (size_t)rem_size : data_size;
     uint32_t i = 0;
 
     // Copy data into current block
@@ -272,7 +272,7 @@ ATCA_STATUS calib_hw_sha2_256_update(ATCADevice device, atca_sha256_ctx_t* ctx, 
     // Save any remaining data
     ctx->total_msg_size += (block_count + 1) * ATCA_SHA256_BLOCK_SIZE;
     ctx->block_size = data_size % ATCA_SHA256_BLOCK_SIZE;
-    memcpy(ctx->block, &data[copy_size + block_count * ATCA_SHA256_BLOCK_SIZE], ctx->block_size);
+    memcpy(ctx->block, &data[copy_size + block_count * ATCA_SHA256_BLOCK_SIZE], (size_t)ctx->block_size);
 
     return ATCA_SUCCESS;
 }
@@ -315,7 +315,7 @@ ATCA_STATUS calib_hw_sha2_256_finish(ATCADevice device, atca_sha256_ctx_t* ctx, 
         ctx->block[ctx->block_size++] = 0x80;
 
         // Add padding zeros plus upper 4 bytes of total message size in bits (only supporting 32bit message bit counts)
-        memset(&ctx->block[ctx->block_size], 0, pad_zero_count + 4);
+        memset(&ctx->block[ctx->block_size], 0, (size_t)pad_zero_count + 4);
         ctx->block_size += pad_zero_count + 4;
 
         // Add the total message size in bits to the end of the current block. Technically this is
@@ -342,7 +342,7 @@ ATCA_STATUS calib_hw_sha2_256_finish(ATCADevice device, atca_sha256_ctx_t* ctx, 
     }
     else
     {
-        if (ATCA_SUCCESS != (status = calib_sha_end(device, digest, ctx->block_size, ctx->block)))
+        if (ATCA_SUCCESS != (status = calib_sha_end(device, digest, (uint16_t)ctx->block_size, ctx->block)))
         {
             return ATCA_TRACE(status, "calib_sha_end - failed");
         }
@@ -425,7 +425,7 @@ ATCA_STATUS calib_sha_hmac_update(ATCADevice device, atca_hmac_sha256_ctx_t* ctx
     ATCA_STATUS status = ATCA_SUCCESS;
     uint32_t block_count;
     uint32_t rem_size = ATCA_SHA256_BLOCK_SIZE - ctx->block_size;
-    uint32_t copy_size = data_size > rem_size ? rem_size : (uint32_t)data_size;
+    size_t copy_size = data_size > rem_size ? (size_t)rem_size : data_size;
     uint32_t i = 0;
 
     // Copy data into current block
@@ -458,7 +458,7 @@ ATCA_STATUS calib_sha_hmac_update(ATCADevice device, atca_hmac_sha256_ctx_t* ctx
     // Save any remaining data
     ctx->total_msg_size += (block_count + 1) * ATCA_SHA256_BLOCK_SIZE;
     ctx->block_size = data_size % ATCA_SHA256_BLOCK_SIZE;
-    memcpy(ctx->block, &data[copy_size + block_count * ATCA_SHA256_BLOCK_SIZE], ctx->block_size);
+    memcpy(ctx->block, &data[copy_size + block_count * ATCA_SHA256_BLOCK_SIZE], (size_t)ctx->block_size);
 
     return ATCA_SUCCESS;
 }
@@ -502,7 +502,7 @@ ATCA_STATUS calib_sha_hmac_finish(ATCADevice device, atca_hmac_sha256_ctx_t *ctx
 
     mode |= target;
 
-    return calib_sha_base(device, mode, ctx->block_size, ctx->block, digest, &digest_size);
+    return calib_sha_base(device, mode, (uint16_t)ctx->block_size, ctx->block, digest, &digest_size);
 }
 
 /** \brief Use the SHA command to compute an HMAC/SHA-256 operation.
